requirements.txt
Django==2.2
django-crispy-forms==1.7.2
django-cruds-adminlte==0.0.14
djangoajax==3.1
pkg-resources==0.0.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
